% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brif.R
\docType{package}
\name{brif-package}
\alias{brif-package}
\title{brif: A tree and forest tool for classification and regression}
\description{
Build decision trees and random forests for classification and regression. The implementation strikes a balance between minimizing computing efforts and maximizing the expected predictive accuracy, thus scales well to large data sets. Multi-threading is available through 'OpenMP'.
}
\section{Available functions}{

Use \code{\link[brif]{brif}} to build a random forest and (optionally) make predictions.
Use \code{\link[brif]{brifTree}} to build a single decision tree.
Use \code{\link[brif]{printRules}} to print out the decision rules of a tree.
Use \code{\link[brif]{predict.brif}} to make predictions using a brif model (tree or forest).
}

\author{
Yanchao Liu
}
