% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brif.R
\name{brif_write_data}
\alias{brif_write_data}
\title{Write data set to file}
\usage{
brif_write_data(df, resp_col_num = 1, outfile = "data")
}
\arguments{
\item{df}{a data frame}

\item{resp_col_num}{an integer indicating the column number (in df) of the response variable. For test data without the response column, use 0 here.}

\item{outfile}{a character string specifying the file name prefix of output files}
}
\value{
a list of four elements. n: number of rows, p: number of predictors, data_file: name of the data file, config_file: name of the configuration file
}
\description{
Write data set to file
}
