% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printBrifTree}
\alias{printBrifTree}
\title{Print the decision rules of a Brif tree}
\usage{
printBrifTree(rf, which_tree)
}
\arguments{
\item{rf}{an object of class 'brif', as returned by rftrain.}

\item{which_tree}{an integer indicating the tree number}
}
\value{
No return value. The function is intended for producing a side effect, which prints the decision rules to the standard output.
}
\description{
Print the decision rules of a Brif tree
}
