% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rfpredict}
\alias{rfpredict}
\title{Predict new cases}
\usage{
rfpredict(rf, rdf, vote_method, nthreads)
}
\arguments{
\item{rf}{an object of class 'brif', as returned by rftrain.}

\item{rdf}{a data frame containing the new cases to be predicted.}

\item{vote_method}{an integer (0 or 1) indicating the voting mechanism among leaf predictions.}

\item{nthreads}{an integer specifying the number of threads to be used in prediction.}
}
\value{
a data frame containing the predicted values.
}
\description{
This function is not intended for end users. Users should use the predict.brif function instead.
}
