% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rftrain}
\alias{rftrain}
\title{Train a random forest}
\usage{
rftrain(rdf, par)
}
\arguments{
\item{rdf}{a data frame. The first column is treated as the target variable.}

\item{par}{a list containing all parameters.}
}
\value{
a list, of class "brif", containing the trained random forest model.
}
\description{
This function is not intended for end users. Users should use the brif.formula or brif.default function.
}
