% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rftrainpredict}
\alias{rftrainpredict}
\title{Train a model and predict for newdata in one go}
\usage{
rftrainpredict(rdf, rdf_new, par)
}
\arguments{
\item{rdf}{a data frame containing the training data.}

\item{rdf_new}{a data frame containing new cases to be predicted.}

\item{par}{a list containing all parameters.}
}
\value{
a data frame containing the predicted values.
}
\description{
This function is not intended for end users. Users should use the function brif or brif.trainpredict and supply the newdata argument thereof.
}
