% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brif.R
\name{stratpar}
\alias{stratpar}
\title{Stratified permutation of rows by the first column}
\usage{
stratpar(x, stride)
}
\arguments{
\item{x}{a data frame to be permuted by row}

\item{stride}{an integer indicating how many rows are to be groups in one block}
}
\value{
a data frame, which is a permutation of x
}
\description{
Stratified permutation of rows by the first column
}
